Add rule for header comments
This adds a rule to warn when files have copyright boilerplate comments.
This depends on https://github.com/realm/SwiftLint/pull/124 for the
tests to run successfully. There's also the issue that the integration
tests for SwiftLint don't work with this. Maybe this would need to wait
for the new defaults to disable this rule by default.