add opt-in IBOutlet Setter Rule
This rule checks if the file contains IBOutlets, if it does it looks for any
lines of code that attempt to call the setter on that outlet.
e.g. self.myLabel = UILabel() would trigger this rule.

While the rule is not perfect, it will help avoid an Interface Builder
anti- pattern. In general, if your code is resetting IBOutlets, there is
probably a better way to accomplish what you want.

There are a lot of triggering and non-triggering test cases. Along with a 3
TODOs where the rule can be triggered when it shouldn't. Because this is opt-in
and the cases where the rule is triggered are fairly rare and can often be fixed
by renaming a local variable, I left them as TODOs for now.