BodyLengthConfiguration for length rules allowing exclusions
BodyLengthConfiguration is a new configuration for the length rules (file length, type body length and function body length).

Besides the already existing warning and error configuration options, it allows a third one which applies exclusions to the specific rule. That way, one can keep the rule enabled and with sane defaults, but avoid the rule triggering in a common set of file, types or functions. The main example will be XCTestCase subtypes which might go for more than 400 lines and need disabling comments in each of the files.