Introduction of a new rule
I have attempted to introduce a new rule for Swiftlint. The rule is designed to warn against violations of the order of keywords upon defining access control and overriding on any type of property or function.
Using Xcode's autocompletion, will result the definition of the access control to be put first, and the override keyword second. My dream is to enforce this rule for every Swift project.