Pass storage into visitLintableFiles block instead of capturing it
visitLintableFiles was given an inout RuleStorage to collect rules into,
but was also given a closure which captured the same RuleStorage
instance. When this closure was called, Swift's exclusivity checker
would trap, since the value was being read by the closure while it was
being mutated by visitLintableFiles.

Passing the RuleStorage into the closure at the call site fixes this by
guaranteeing that the storage given to the closure isn't being
simultaneously mutated on the thread.