Add optional support for Missing Import Result
Swift 5 and 5.1 compiled code (including Swift 4, 4.1, 4.2 code compiled by Swift 5) all have access to the intrinsic Result type. If your code still depends on the `Result` (v4.0.1 and earlier) framework, then you must import it or your code will either not compile or silently compile with the intrinsic `Result` type. Turn this rule on in a project that depends on this framework to avoid missing this crucial import. NB: Version 5.0 of `Result` makes it compatible with Swift 5's `Result` type.