Add IBInspectableInExtensionRule
This rule validates that there are no `@IBInspectable` variables in
`extension`s. The reason for this is that Interface Builder doesn't take
into account what code is actually included in your target to determine
if an inspectable should be valid. So in this case:

```swift
// Module A
extension UIView {
  @IBInspectable public var foo: String? {
    // do something
  }
}
```

```swift
// Module B
// Uses `foo` in IB
```

If module B does not depend on module A interface builder will still
show the `foo` IBInspectable, but it will fail at runtime.