Fix UnusedImportRule breaking transitive imports
Previously in the case a unused import was allowed to transitively
pollute another module, when that unused import was removed, this rule
would not insert the previously transitively included module, leading
the code to fail to compile. Now unused imports, that will be removed,
are no longer considered for which imports can be implicit.