Add instructions for releasing a fork
While I was working on a patch for SwiftLint, I needed to be able to run it from the root of my project
in order to reproduce the bug. It took a lot of trial and error to figure out how to get my project to pick
up my fork's binaries. It'd be great to have these instructions as part of the repo to help future
contributors.