Allow multiline parameters rule to always break lines
When I originally implemented this rule, I made a design choice to allow either all the parameter in one line or in multiple lines, but there are cases/teams which prefer to have all the parameters in different lines.

This PR adds the support of this use case. It's nows possible to configure the Multiline Parameters Rule to enforce breaking the list of parameters into multiple lines.