Introduce a model for rule configuration description
The idea is to not just build a description of a rule configuration by
crafting a string representation manually but applying a defined set
of elements. These elements are put together by a new result builder.
The model enables a common but different formatting of the console
output as well as the representation on the rule's web page. Predefined
types make sure that it is clear whether to put the values in quotes or
brackets. Nested configurations are also considered.