Make comma_inheritance opt-in
The Swift language grammar doesn't allow a comma in place of an ampersand in all cases. Specifically, if a protocol's associated type is required to conform to a composition of protocols and one of those protocols in turn requires conformance to one of the same protocols, the compiler will generate an error: "redundant conformance constraint". For example:

    protocol Transformable {
        associatedtype Unit: Equatable & CaseIterable & UnitConvertible

        ...
    }

    protocol UnitConvertible where Self: Equatable {
        ...
    }

If comma_inheritance is enabled by default, autocorrection can introduce a new compile error. Making the rule opt-in allows code bases where it isn't a problem to opt-in and continue to use autocorrect, while code bases where it is a problem can leave it disabled.