Run rules serially
When rules were mostly SourceKit-based, linting all rules concurrently
on any given file had a performance boost.

Now that most rules are SwiftSyntax-based, it's possible this is no
longer helping and in fact may be increasing thread/core contention or
otherwise adding overhead.

Best way to know is to measure, so I'm doing that here.