Add `dead-code` command
This command should be considered experimental because there's only
partial integration with the rest of SwiftLint's infrastructure, and the
command is very likely to change in the near future.

This can effectively replace the `unused_declaration` analyzer rule and
in fact re-uses the same rule identifier to simplify migrations.

However, instead of using SourceKit requests, this parses the index
store generated by the Swift compiler, in addition to SwiftSyntax, which
means it runs it can run in seconds instead of hours when analyzing
projects with tens of thousands of Swift files.

In addition to Swift, this also catches dead C, C++, Objective-C
declarations, mostly by accident. There's no way to disable the rule for
those files at the moment.