Remove the need for rules_xcodeproj dependency for downstream consumers
This was needed because the `xcodeproj` target was declared in the same
BUILD file with the `@SwiftLint//:swiftlint` target, so when users run
`bazel run @SwiftLint//:swiftlint`, bazel would try to load
`rules_xcodeproj`. `rules_xcodeproj` 1.0 allows us to specify the
location of the generated Xcode project, so we don't need this target in
the top-level BUILD file anymore.

This only fixes for non-bzlmod case now.

Fixes https://github.com/realm/SwiftLint/issues/4737