Add no_default_switch_in_operators rule
This rule validates that `default` is not used in switch cases in
operator definitions. This could be expanded to do more than operator
definitions but operators feel a bit special in that forgetting to add
your case to them as you add new cases probably leads to more subtle
bugs than other areas.