#5205 Explicitely disable cache for SwiftLintPlugin runs on Xcode Cloud
In Xcode Cloud environment, SwiftLint’s cache cannot be written. When using the SwiftLinkPlugin, there is no way to disable the cache. Previously, a solution was made for the SwiftLint CLI itself where it looks at a set of environment variables (#4485). This solution offers a cleaner approach where the plugin itself decides whether it needs to enable or disable the cache based on the `CI_XCODE_CLOUD` environment variable.