Prevent concatenation of SwiftUI text elements (#3)
We want to prevent developers from doing things that break translations
such as concatenating two `Text` elements in order to apply different
styling to each. The proper way to do it is to use `AttributedString`
and apply different styling that way.

This adds a rule to look for binary `+` operators with `Text.init` calls
on both sides.