RedundantTypeAnnotation: add 'ignore_type_interfaces' option
The `redundant_type_annotation` and `explicit_type_interface` rules conflict.
For users that want to have an explicit type interface, but still be able to write
simple code inside the bodies of functions, I want to add an option to
`redundant_type_annotation` that has it ignore "type interfaces".

For example, this allows:
```
struct Foo {
    var bar: Bar = Bar() // OK: ignore this! I want explicit types on interfaces.
    func baz() {
       let bar: Bar = Bar() // WARN: redundant_type_annotation kicks in here though.
    }
}
```