Separate Docker build into artifact and release version
This builds the Docker image twice per release - first to extract the
Linux binary from it as a dedicated release artifact, second as the
official Docker image associated with the later published release.

This avoids that a Docker image for a version exist that is not otherwise
available as an official release on GitHub.