Revert back to building Linux binaries with Docker
The binaries built with Bazel are not statically linked. That means
that they require a different set of shared libraries to be available
at runtime. This would be a breaking change. As an alternative, would
could find out how to statically link in the Bazel build or build with
SPM.