Introduce mimalloc to static Linux binaries (#6321)
mimalloc is performant memory allocator written in C that works as drop-in replacement for malloc and supports builds for musl. Compared with other popular allocators like jemalloc or tcmalloc, it's a fair option. 

https://github.com/microsoft/mimalloc

Also, in the Swift compiler project (swiftlang/swift), mimaloc has been adopted as the allocator in Windows's toolchain (not runtime!):

https://forums.swift.org/t/se-454-adopt-mimalloc-for-windows-toolchain/

Co-authored-by: Danny Mösch <danny.moesch@icloud.com>