Tests: adjust flag computation for Windows
Windows isolates XCTest/Testing from the platform SDK as they are not
part of the redistributable components. We adjust the flags by querying
the partial path information and computing the location for the `XCTest`
module for the tests.

Additionally, enable the ObjC interop on the test cases to allow parsing
`@objc` modifiers on declarations. This allows us to pass the
GeneratedTests test suite.