correct incorrect handling of file paths
`URL.path` is not a usable rendering of the path. `URL.path` represents
the URI path. If the path is meant to be consumable as a file path, the
file system representation (aka FSR) must be retrieved.

Unfortunately, the FSR access is a bit more complicated as the API for
that returns a +0 borrowed object with the lifetime bound to `URL`.
Because this relies on the toll-free bridged storage of `URL`, Swift
does not have this. Instead, we are provided a
`withUnsafeFileSystemRepresentation` which provides a heap allocated
buffer that is scoped, which we can copy into a `String`. To abstract
away this wart, we introduce an extension on `URL` to access the FSR
through the new `filepath` property.

This improves the file path handling and makes additional tests now
pass.